+++ waimak-core/src/main/scala/com/coxautodata/waimak/dataflow/DataFlow.scala
@@ -472,12 +472,6 @@ object DataFlow {
 
   val dataFlowParamPrefix: String = "spark.waimak.dataflow"
 
-  /**
-    * Whether to cache labels before they are committed if they are reused
-    * elsewhere in the flow.
-    */
-  val CACHE_REUSED_COMMITTED_LABELS: String = s"$dataFlowParamPrefix.cacheReusedCommittedLabels"
-  val CACHE_REUSED_COMMITTED_LABELS_DEFAULT: Boolean = true
 }
 
 /**
