+++ waimak-core/src/main/scala/com/coxautodata/waimak/dataflow/package.scala
@@ -1,6 +1,6 @@
 package com.coxautodata.waimak
 
-import com.coxautodata.waimak.dataflow.DataFlow.{CACHE_REUSED_COMMITTED_LABELS, CACHE_REUSED_COMMITTED_LABELS_DEFAULT}
+import com.coxautodata.waimak.dataflow.CommitExtension.{CACHE_REUSED_COMMITTED_LABELS, CACHE_REUSED_COMMITTED_LABELS_DEFAULT}
 
 /**
   * Created by Alexei Perelighin on 2018/01/11.
