+++ waimak-core/src/main/scala/com/coxautodata/waimak/dataflow/spark/package.scala
@@ -636,7 +636,7 @@ package object spark {
     def cacheAsParquet(labels: String*): SparkDataFlow = {
       if (labels.isEmpty) throw new DataFlowException(s"At least one label must be specified for cacheAsParquet")
 
-      labels.foldLeft(sparkDataFlow) { (flow, label) => SparkInterceptors.addCacheAsParquet(flow, label, None, repartition = false) }
+      labels.foldLeft(sparkDataFlow) { (flow, label) => CacheAsParquetExtension.addCacheAsParquet(flow, label, None, repartition = false) }
     }
 
     /**
@@ -651,7 +651,7 @@ package object spark {
     def cacheAsPartitionedParquet(partitions: Seq[String], repartition: Boolean = true)(labels: String*): SparkDataFlow = {
       if (labels.isEmpty) throw new DataFlowException(s"At least one label must be specified for cacheAsParquet")
 
-      labels.foldLeft(sparkDataFlow) { (flow, label) => SparkInterceptors.addCacheAsParquet(flow, label, Some(Left(partitions)), repartition) }
+      labels.foldLeft(sparkDataFlow) { (flow, label) => CacheAsParquetExtension.addCacheAsParquet(flow, label, Some(Left(partitions)), repartition) }
     }
 
     /**
